requirements.txt
load-dotenv==0.1.0
psycopg2-binary==2.9.10
pyspark==3.5.4
python-dateutil==2.9.0.post0
six==1.17.0
typing-extensions==4.12.2
